Replace Glue packages with IOConnect packages for react/solutions tutorial
