--- react/solution/client-details/src/glue.js
@@ -1,9 +1,0 @@
-export const setClientFromWorkspace = (setClient) => async (glue) => {
-    const myWorkspace = await glue.workspaces.getMyWorkspace();
-    myWorkspace.onContextUpdated((context) => {
-        if (context) {
-            setClient(context);
-            myWorkspace.setTitle(context.clientName);
-        };
-    });
-};
