+++ react/solution/clients/package.json
@@ -3,8 +3,8 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
-        "@glue42/react-hooks": "^2.8.3",
-        "@glue42/workspaces-api": "^1.17.1",
+        "@interopio/react-hooks": "^3.5.2",
+        "@interopio/workspaces-api": "^3.5.2",
         "@testing-library/jest-dom": "^4.2.4",
         "@testing-library/react": "^9.3.2",
         "@testing-library/user-event": "^7.1.2",
@@ -17,7 +17,7 @@
         "react-select": "^3.1.0"
     },
     "scripts": {
-        "start": "react-scripts start",
+        "start": "cross-env PORT=3000 react-scripts start",
         "build": "react-scripts build",
         "test": "react-scripts test",
         "eject": "react-scripts eject"
@@ -37,5 +37,8 @@
             "last 1 firefox version",
             "last 1 safari version"
         ]
+    },
+    "devDependencies": {
+        "cross-env": "^7.0.3"
     }
 }
