+++ react/solution/clients/public/index.html
@@ -22,6 +22,7 @@
     <body>
         <noscript>You need to enable JavaScript to run this app.</noscript>
         <div id="root"></div>
+        <script type="module" src="/src/index.jsx" ></script>
         <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
