+++ react/solution/portfolio-downloader/package.json
@@ -4,11 +4,10 @@
     "homepage": "/portfolio-downloader/",
     "private": true,
     "dependencies": {
-        "@glue42/react-hooks": "^2.8.3",
-        "@glue42/web": "^2.8.3",
-        "@testing-library/jest-dom": "^5.16.5",
-        "@testing-library/react": "^13.4.0",
-        "@testing-library/user-event": "^13.5.0",
+        "@interopio/react-hooks": "^3.5.2",
+        "@testing-library/jest-dom": "^5.11.4",
+        "@testing-library/react": "^11.1.0",
+        "@testing-library/user-event": "^12.1.10",
         "bootstrap": "^4.4.1",
         "react": "^16.13.0",
         "react-dom": "^16.13.0",
@@ -16,7 +15,7 @@
         "web-vitals": "^2.1.4"
     },
     "scripts": {
-        "start": "react-scripts start",
+        "start": "cross-env PORT=9400 react-scripts start",
         "build": "react-scripts build",
         "test": "react-scripts test",
         "eject": "react-scripts eject"
@@ -38,5 +37,8 @@
             "last 1 firefox version",
             "last 1 safari version"
         ]
+    },
+    "devDependencies": {
+        "cross-env": "^7.0.3"
     }
 }
