+++ react/solution/stock-details/package.json
@@ -3,7 +3,7 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
-        "@glue42/react-hooks": "^2.8.3",
+        "@interopio/react-hooks": "^3.5.2",
         "@testing-library/jest-dom": "^4.2.4",
         "@testing-library/react": "^9.5.0",
         "@testing-library/user-event": "^7.2.1",
@@ -16,7 +16,7 @@
         "react-select": "^3.1.0"
     },
     "scripts": {
-        "start": "react-scripts start",
+        "start": "cross-env PORT=3002 react-scripts start",
         "build": "react-scripts build",
         "test": "react-scripts test",
         "eject": "react-scripts eject"
@@ -36,5 +36,8 @@
             "last 1 firefox version",
             "last 1 safari version"
         ]
+    },
+    "devDependencies": {
+        "cross-env": "^7.0.3"
     }
 }
