+++ react/solution/stock-details/src/StockDetails.jsx
@@ -1,28 +1,28 @@
 import React, { useContext, useState } from "react";
-import { GlueContext, useGlue } from "@glue42/react-hooks";
-import { getMyWindowContext, subscribeForInstrumentStream } from "./glue";
+import {useIOConnect, IOConnectContext} from "@interopio/react-hooks"
+import { getMyWindowContext, subscribeForInstrumentStream } from "./io";
 
 function StockDetails() {
-    const glue = useContext(GlueContext);
+    const io = useContext(IOConnectContext);
     const [windowContext, setWindowContext] = useState({});
-    useGlue(getMyWindowContext(setWindowContext));
+    useIOConnect(getMyWindowContext(setWindowContext));
     const {
         stock: { RIC, BPOD, Bloomberg, Description, Exchange, Venues } = {}
     } = windowContext || {};
     const [{ Bid, Ask }, setPrices] = useState({ Bid: windowContext.Bid, Ask: windowContext.Ask});
-    useGlue(subscribeForInstrumentStream(setPrices), [RIC]);
+    useIOConnect(subscribeForInstrumentStream(setPrices), [RIC]);
 
     return (
         <div className="container-fluid">
             <div className="row">
                 <div className="col-md-2">
-                    {!glue && (
-                        <span id="glueSpan" className="badge badge-warning">
+                    {!io && (
+                        <span id="ioSpan" className="badge badge-warning">
                             io.Connect is unavailable
                         </span>
                     )}
-                    {glue && (
-                        <span id="glueSpan" className="badge badge-success">
+                    {io && (
+                        <span id="ioSpan" className="badge badge-success">
                             io.Connect is available
                         </span>
                     )}
