+++ react/solution/stocks/src/Stocks.jsx
@@ -1,22 +1,22 @@
 import React, { useEffect, useState, useContext } from "react";
 import { REQUEST_OPTIONS } from "./constants";
-import { GlueContext, useGlue } from "@glue42/react-hooks";
+import { useIOConnect, IOConnectContext } from "@interopio/react-hooks";
 import {
     createInstrumentStream,
     subscribeForInstrumentStream,
     setClientFromWorkspace,
     openStockDetailsInWorkspace,
     raiseExportPortfolioIntentRequest
-} from "./glue";
+} from "./io";
 
 function Stocks() {
     const [portfolio, setPortfolio] = useState([]);
     const [{ clientId, clientName }, setClient] = useState({});
     const [prices, setPrices] = useState({});
-    const subscription = useGlue(
-        (glue, portfolio) => {
+    const subscription = useIOConnect(
+        (io, portfolio) => {
             if (portfolio.length > 0) {
-                return subscribeForInstrumentStream(setPrices)(glue, portfolio);
+                return subscribeForInstrumentStream(setPrices)(io, portfolio);
             }
         },
         [portfolio]
@@ -40,24 +40,24 @@ function Stocks() {
         fetchPortfolio();
     }, [clientId]);
 
-    const glue = useContext(GlueContext);
-    const showStockDetails = useGlue(openStockDetailsInWorkspace);
-    useGlue(createInstrumentStream);
+    const io = useContext(IOConnectContext);
+    const showStockDetails = useIOConnect(openStockDetailsInWorkspace);
+    useIOConnect(createInstrumentStream);
     const setDefaultClient = () => setClient({ clientId: "", clientName: "" });
-    useGlue(setClientFromWorkspace(setClient));
-    const exportPortfolioButtonHandler = useGlue(raiseExportPortfolioIntentRequest);
+    useIOConnect(setClientFromWorkspace(setClient));
+    const exportPortfolioButtonHandler = useIOConnect(raiseExportPortfolioIntentRequest);
 
     return (
         <div className="container-fluid">
             <div className="row">
                 <div className="col-md-2">
-                    {!glue && (
-                        <span id="glueSpan" className="badge badge-warning">
+                    {!io && (
+                        <span id="ioSpan" className="badge badge-warning">
                             io.Connect is unavailable
                         </span>
                     )}
-                    {glue && (
-                        <span id="glueSpan" className="badge badge-success">
+                    {io && (
+                        <span id="ioSpan" className="badge badge-success">
                             io.Connect is available
                         </span>
                     )}
